renaming file names to snakecase
