+++ requirements.txt
@@ -17,7 +17,7 @@ xmltodict
 jsonschema==2.6.0
 
 # cx_Freeze is used for packaging, the pypi version does not have a fix for Python 3.7 on Windows 10 (https://github.com/anthony-tuininga/cx_Freeze/issues/407)
-git+https://github.com/anthony-tuininga/cx_Freeze.git@master
+# git+https://github.com/anthony-tuininga/cx_Freeze.git@master
 
 # PySide2 has had some issues with new versions and packaging so we will pin it
 PySide2==5.12.3
